refactor: #8 Categoryの型定義をmodelファイルに移行。
