+++ bear-clone/src/app/shared/models/category.model.ts
@@ -1,0 +1,8 @@
+import { IconDefinition } from "@fortawesome/fontawesome-svg-core";
+
+export type Category = {
+  name: string;
+  icon: IconDefinition;
+  isOpen: boolean;
+  categories?: Category[];
+};
